Admin user changes

+++ app/Http/Controllers/UserController.php
@@ -40,31 +40,51 @@ class UserController extends Controller
 
     // Delete the any user.
     public function destroy($id)
-    {
-
-        $hasEvents = Events::where('organizer_id', $id)->first();
-        if ($hasEvents) {
-            return redirect()->back()->with('error', 'The user has organized events and cannot be deleted.');
-        }
-        $hasOrders = Order::where('user_id', $id)->first();
-        if ($hasOrders) {
-            return redirect()->back()->with('error', 'The user has orders as an organizer and cannot be deleted.');
-        } else {
-            $user = User::findOrFail($id);
-            $user->delete();
-            // return response()->json(['su' => 'User deleted successfully'])
-            return redirect()->back()->with('success', 'User Deleted Succefully.');
+    { {
+            $hasEvents = Events::where('organizer_id', $id)->first();
+            if ($hasEvents) {
+                return redirect()->back()->with('error', 'The user has organized events and cannot be deleted.');
+            }
+            $hasOrders = Order::where('user_id', $id)->first();
+            if ($hasOrders) {
+                return redirect()->back()->with('error', 'The user has orders as an organizer and cannot be deleted.');
+            }
+            $donePayment = User::whereNotNull('email_verified_at');
+            if ($donePayment) {
+                return redirect()->back()->with('error', 'You can not delete this account because it has already paid!');
+            } else {
+                $user = User::findOrFail($id);
+                $user->delete();
+                return redirect()->back()->with('success', 'User Deleted Succefully.');
+            }
         }
     }
-    // Show user the data of the purchased tickets by the users
     public function purchasedBy($id)
     {
         $organizer = Auth::user();
         $event = Events::find($id);
-        //  here other than the organizer admin can also access this page. if the organizer user is not valid than it will show error page.
         if ($organizer && $event && $organizer->id === 0 || $event->organizer_id === $organizer->id) {
             $purchases = Order::where('event_id', $id)->with('user')->get();
             return view('Admin.purchasedBy', compact('event', 'purchases'));
+        } else {
+            abort(403, 'Unauthorized');
+        }
+    }
+
+    public function UserStatistics()
+    {
+        $user_id =  Auth::user()->id;
+        if ($user_id == 0) {
+            $paidUsers = User::whereNotNull('email_verified_at')->paginate(10);
+            $totalPaidUsers = User::whereNotNull('email_verified_at')->count();
+            $unverifiedUsers = User::whereNull('email_verified_at')->get();
+
+
+            return view('Admin.UserStatistics', [
+                'paidUsers' => $paidUsers,
+                'unverifiedUsers' => $unverifiedUsers,
+                'totalPaidUsers' => $totalPaidUsers,
+            ]);
         } else {
             abort(403, 'Unauthorized');
         }
